SSTTP-1162 - Wrong interest amount shown initially upon amending debts
When an amount was being removed, the calculator input was being updated
in keystore but the schedule was not. As a result, when the calculator page
was loading, it was matching against Some(schedule) and loading the old data.
Now, when an amount is removed, schedule is set to None to force the calculator
to re-calculate when it initially loads.